First Rate Limiting attempt.
The implementation tests after applying policies to prevent having to load revisions after every change made by the Aggregator itself, at the cost of limiting *after* the value shave been calculated.

Not entirely happy about the fact that IWorkItem needs to be aware of the concept of rate limiting, but there is presently no easy way to expose the Revisions outside of the WorkItemWrapper class.

When ChangedDate of the previous save > configured limit, no revisions are loaded.

Should fix #20.