Fix issue #160
Each request get a new RuntimeContext clone, but the ScriptEngine was created at later time and stored in the clone... which is discarded at the end of the request. Net result: no caching of the scripting engine.

Fixed by storing the computed engine in the cache. It cannot be built at the same time as the context because we do not have the WorkitemRepo, which is built later.